Authentication & Authorization

+++ App/controllers/ListingController.php
@@ -2,7 +2,10 @@
 
 namespace App\Controllers;
 
+use Framework\Authorization;
 use Framework\Database;
+use Framework\Middleware\Authorize;
+use Framework\Session;
 use Framework\Validation;
 
 class ListingController
@@ -24,7 +27,7 @@ class ListingController
   public function index()
   {
 
-    $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+    $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
     loadView('listings/index', [
       'listings' => $listings
     ]);
@@ -77,7 +80,7 @@ class ListingController
 
     $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-    $newListingData['user_id'] = 1;
+    $newListingData['user_id'] = Session::get('user')['id'];
 
     $newListingData = array_map('sanitize', $newListingData);
 
@@ -124,8 +127,8 @@ class ListingController
 
       $this->db->query($query, $newListingData);
 
-      // Add this line
-      $_SESSION['success_message'] = 'Listing created successfully';
+      // Set Session Message
+      Session::setFlashMessage('success_message', 'Listing created successfully');
 
       redirect('/listings');
     }
@@ -152,10 +155,17 @@ class ListingController
       ErrorController::notFound('Listing not found');
       return;
     }
+
+    //Authorization
+    if (!Authorization::isOwner($listing->user_id)) {
+      Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+      return redirect('/listings/' . $id);
+    }
+
     $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
     //Set flash message
-    $_SESSION['success_message'] = 'Listing deleted successfully';
+    Session::setFlashMessage('success_message', 'Listing deleted successfully');
 
     redirect('/listings');
   }
@@ -179,6 +189,12 @@ class ListingController
     if (!$listing) {
       ErrorController::notFound('Listing not found');
       return;
+    }
+
+    //Authorization
+    if (!Authorization::isOwner($listing->user_id)) {
+      Session::setFlashMessage('error_message', 'You are not authorized to edit this listing');
+      return redirect('/listings/' . $id);
     }
 
     loadView('listings/edit', [
@@ -205,6 +221,12 @@ class ListingController
     if (!$listing) {
       ErrorController::notFound('Listing not found');
       return;
+    }
+
+    //Authorization
+    if (!Authorization::isOwner($listing->user_id)) {
+      Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+      return redirect('/listings');
     }
 
     $allowedFields = ['title', 'description', 'salary', 'tags', 'company', 'address', 'city', 'state', 'phone', 'email', 'requirements', 'benefits'];
@@ -247,9 +269,35 @@ class ListingController
 
       $this->db->query($updateQuery, $updateValues);
 
-      $_SESSION['success_message'] = 'Listing Updated';
+      Session::setFlashMessage('success_message', 'Listing updated');
 
       redirect('/listings/' . $id);
     }
+  }
+
+  /**
+   * Search listings by keywords/location
+   * 
+   * @return void
+   */
+  public function search()
+  {
+    $keywords = isset($_GET['keywords']) ? trim($_GET['keywords']) : '';
+    $location = isset($_GET['location']) ? trim($_GET['location']) : '';
+
+    $query = "SELECT * FROM listings WHERE (title LIKE :keywords OR description LIKE :keywords OR tags LIKE :keywords OR company LIKE :keywords) AND (city LIKE :location OR state LIKE :location)";
+
+    $params = [
+      'keywords' => "%{$keywords}%",
+      'location' => "%{$location}%",
+    ];
+
+    $listings = $this->db->query($query, $params)->fetchAll();
+
+    loadView('/listings/index', [
+      'listings' => $listings,
+      'keywords' => $keywords,
+      'location' => $location,
+    ]);
   }
 }
